Rename space_ to semi_space_. (#520)
* Adapt CloneInToSpace from Dartino to Toit.

* Merge Heap and TwoSpaceHeap.

We only have two-space heaps so no need to
have this difference with lots of virtual
methods.  Merges the classes, but doesn't
move code around yet because that would make
the CL unreadable.  (There's still a non-Dartino
class called Heap that needs to be merged.)

* feedback

* Rename space_ to semi_space_.

Also rearranges TwoSpaceHeap so it has a more
conventional public-protected-private ordering.

* New files won't link on Windows yet