Allow parallel init of WiFi module. (#572)
When two processes are trying to connect to the WiFi at the same time,
then the second one falls back to using the System network.
However, at that time the network might not yet be online. We now wait
for it to be online by waiting for the IP-address to be available.