Merge Heap and ObjectHeap classes into one. (#593)
This also means their constructors, destructors and
_expand methods are merged.  In this CL we don't reorder
the .h file, so private and public parts are not cleanly
separated.  That will be done in a subsequent moving-only
CL.

Also takes iterate_roots out of the gc() function and
makes it an independent method on ObjectHeap.

Removes Heap::payload_size() which was unused.

Moves NoGC class.