Allocate initial chunks for new GC heap early. (#606)
* Allocate initial chunks for new GC heap early.

A new heap must immediately have some space to
allocate, since some allocations happen immediately
and can't trigger a GC.

But in Toit any allocation can fail, and that includes the
initial pages for a new heap for a new process.  We handle
this by allocating the first page in the primitive before
trying to create the new process.  This way we can report
back allocation failures to the primitive machinery that
retries the primitive.

This CL extends that logic to the new GC.  In the case of the
new GC we start each two-space heap off with a page per
semispace, which means we will be able to perform a young
space GC.

* nit