Avoid creating an old-space if it is not needed.
If you use set_max_heap_size_ to set the max heap
to a single page (4k on embedded, 32k on desktop/server)
then no old space will be created, and the heap of the
process will be limited to one page, even if some objects
are long lived.

If the max heap size never exceeds a half page (2k or 16k)
then we also create no old-space.