Improve stacktrace.toit. (#803)
This tool (for native stacktraces) now uses host.arguments
to parse command line args.

Also improves it so it can use any elf-compatible objdump if
the Xtensa objdump is not installed.  Demangling may or may
not work in this case.

Faster for the the non-disassembly case where there is no
need to parse millions of lines of disassembly.

By using the new version of pkg-host we detect if objdump
exits with an error without producing meaningful output.