Fixes bad resolution of index-calls with parameters
Fixes a confusion in the resolver.
Take the call `o[3] 4`

The parser would correctly parse this as a call of `o[3]` with argument
`4`.

However, the resolver tried to reuse functionality and gave a
call-resolving function just a "target" and "arguments".

For a normal call the "target" was the target of the call (for `foo 1
2`, the `foo` identifier).
For an index call, the "target" was the node itself: `o[3]`.

Once the target and arguments were resolved, the resolver would then
again check whether it was trying to solve a normal call or an index
call. It did this, by looking at the target-node. If it was and
index-node, it assumed that it was resolving and index-call. Otherwise,
it assumed it was for a call-node.

However, this led to the following bug, where a normal node had an
index-call as target. For our example `o[3] 4`, the resolver now got
confused and thought that it was doing work for an index node (since
the target of the call was `o[3]`, and index node). It assumed that
the argument `4` was the argument to the index (as if the user had
written `o[4]`). The actual arguments to the index-node were completely
dropped.