Fix Toit/mbedtls define mismatch.
The Toit code and the mbedtls code did not agree on the defines that
were set. This meant that Toit believed that some of the mbedtls
structures were smaller than they needed to be. The mbedtls code was
writing into fields that were outside the allocated structure,
overwriting neighboring objects.