Provide a lot of standard CRC algorithms.
Deprecates the two we already had in favour of the ones in the common
crc/crc.toit import.  These were CRC-32, in crc/crc32 and and
CRC-16/XMODEM, in crc/crc16.

For each algorithm there is a top level function that returns the
checksum as an int.  There is also a class that conforms to the Checksum
interface by returning a small byte array.  For the big-endian CRC
algorithms, this byte array is in big-endian order, ready to be pasted
into the communications protocols that use it.