Use a special type for stop_bits of uart. (#1173)
It's confusing that --stop_bits=2 runs just fine,
but sets the number of stop bits to 1.5.  To avoid
this, make a special type.  Should be backwards
compatible except in the case where someone used
--stop_bits=1, which coincidentally worked fine.
However this was also the default, so there would
be no need to do this explicitly.