Rewrite 'pin.wait_for'. (#1206)
When calling `wait_for` we now return a "timestamp". Whenever there is
an edge-detection we notify listeners. The listeners can use this
timestamp to know whether the edge-transition happened after they
subscribed or before. If it was before, they know that the requested
level was achieved. The edge-transition might actually indicate that
the pin is now at the non-requested input level, but we still know
that the pin was at least once at the state we are looking for.