When rehashing a Set or Map, reduce the peak memory used.
For large sets or maps we build a new index from scratch
when we outgrow the old one.  In this case we reuse the
arraylets from the old index, rather than creating new
arraylets and keeping the old ones alive during rehashing.

This also changes the rules so that we go to LargeArray_
as soon as we exceed the arraylet size.  Previously we
stayed at SmallArray_ until we hit the max size for page
allocations.  On embedded this will avoid arrays in the
2-3.9k size range, which could waste memory on fragmentation.