Allow fractional seconds in time parsing. (#1544)
It turns out that ISO 8601 is a very lax standard allowing all sorts of
strange date/time formats, inclusing missing parts, eg. 2023-01 for
January 2023. It also allows fractional months, like 2023-01.53743.

Instead of trying to support the whole ISO standard, which would cause
huge bloat in the standard library, we instead support the sane subset,
which is documented in RFC 3339.