Improve tree shaking and fix optimization bug
The tree-grower would grow static calls even if they were to instance
methods of classes that were not instantiated. These static calls were
introduced by an earlier optimization. For example, if there is a static
type check before that guarantees (assuming the check works) only one
dynamic method can be reached.

Now the grower treats these calls as "filtered" virtual calls. That is,
they are selector calls, but if they trigger they can only grow to the
target method. As such, they also have to wait for classes to be
instantiated before they are added to the grown method set.

Care must be taken with super calls: these are also static calls to
instance methods, but we have to handle them as soon as we seen them.

This patch also fixes a bug where a call to `operator ==` was changed
into a static call, even when the RHS was not guaranteed to be non-null.