Zlib primitives and library. (#1722)
This is libs and primitives for zlib, which means programs can use the
deflate algorithm.

It is not currently made available on embedded platforms because it uses
too much
memory and bloats the VM. This may become a config option at some time.

It is not currently available on Windows or 32 bit Linux since the OSs
tend to ship without
those dynamic libraries, so it would prevent the Toit VM from booting.

It works on MacOS and 64 bit Linux. Those platforms ship with the
dynamic library.

The Toit VM already has primitives to support run length encoding in a
format that
is compatible with zlib. Those are available on all platforms. The
classes for that support
have the read and write methods on the same object. In this case I chose
not to do that
because it can cause confusion as to the meaning of the close method.
Therefore the
zlib.Encoder and zlib.Decoder have a separate reader.