Robustify DNS lookup. (#1762)
When there are multiple DNS servers available we
were previously unclear about whether they were
there because they knew about different domains,
or whether they were there in case some servers
were unreachable.

We now follow the way Linux does things, and use
them for redundancy, not because they have different DNS databases. That
simplifies things, because we immediately return a negative answer
instead of
waiting for DNS servers that may not be available.

Previously we were getting test flakiness if one
server failed to respond for a few seconds. Now
we keep retrying for the full timeout (20s). With the previous semantics
this change would have caused very slow answers in the case where a
non-existent domain was looked up and one of the servers was offline.