Change network service resolve to return all IPs.
Previously we only returned the first even though
the type was a list.  This CL previously returned
a random one, which is better than the first from
a load balancing point of view, but caused a fail
in the network service test.  With this change we
return all IPs allowing both the test to pass and
the IP to be randomized.  (https://xkcd.com/276/)