It turns out we must do all things in order.
We must do the Toit finalizers first, and then the
VM finalizers afterwards. So this undoes some of the
changes. The issue is that a we want to be able to do
the VM freeing immediately at the end of the GC, but
that means we must be all done with marking the objects
reachable from the Toit finalizer lambdas before we
look at the Vm finalizers (for external byte arrays etc.).