Weak maps (#1937)
Adds weak maps, which have their values turned to null
if the values are not otherwise reachable in a mark sweep
GC.

These can be used for caching difficult calculations without
causing out-of-memory errors.  At some point we may want
to restrict the zappng of values to mark-sweep-compact or
malloc-failure-caused GCs.

On the C++ side this is implemented by splitting the Toit finalizer
class into two classes with a common super-class for finalizers
with a callback. Only the map is registered, not individual entries.

A single bit on an object's header indicates that it is either weak or
finalizable.
No class can be both, so maps cannot have finalizers.