encoding.yaml (#1961)
Added decode and encode of yaml documents. 
The parser follows the spec closely. The spec defines a PEG-like
grammar, so the implementation is top-down.

Refactored the json code a bit to enable reuse, on the encoding piece.

I want to apologise for the shear amount of source code in this PR, but
handwritten parsers are a bit verbose.