Analyze each URI in its own project. (#2003)
Same as #1972, but for the Go version of the LSP.

Instead of having one shared project root, find the project root for
each file before running the analysis on it.

Still missing:
- we shouldn't report diagnostics for files that are in different
project roots. Otherwise, we could end up with files that have a warning
when they are not opened, and then one when they aren't (depending on
the project root that is used to analyze them).
- if a summary has changed, then all projects that contain that file
should be analyzed again.