Add 'in' and 'out' mixins (#1921)
Objects that support writing should provide an `out` whereas objects
that support reading from them should provide an `int`. A mixin makes it
easy to support them.

It is also possible to just extend the Reader/Writer classes. For some
classes that might be OK. The `io.Buffer` class is one of them. However,
most classes should use the `in` and `out` mixins. This also has the
advantage that we don't need to make `Reader` and `Writer` an interface.