Make a Deque a List. (#2294)
With this change, a Deque is a subclass of the abstract subclass List,
which means it has useful methods like []=, index-of, and index-of
--binary.

In addition, we add methods remove-at and insert-at, which can be used
to insert and remove items from the middle of a list or deque. If the
insertion is near the ends of a deque, or near the end of a List_ then
the operation is efficient.